+++ src/assets/css/order.css
@@ -17,7 +17,14 @@
   width: 100px ;
   justify-content: center;
 }
-
+.excelBtn {
+  width: 100px;
+  height: 26px;
+  border: none;
+  background-color: white;
+  cursor: pointer;
+  margin-right: 0.5%;
+}
 .radio-group label {
   display: flex;
   align-items: center;
