+++ src/components/layouts/header/AdminHeader.vue
@@ -186,33 +186,6 @@ const fetchUsernameFromToken = () => {
 };
 
 const pdfDownload = async () => {
-  if (store.getters.userRole === 'ROLE_ROOT') {
-    try {
-      const response = await fetch('http://localhost:5000/admin/pdfdownload/root-pdf', {
-        method: 'GET',
-        headers: {
-          'Authorization': `Bearer ${accessToken}`,
-          'Content-Type': 'application/json',
-        },
-      });
-
-      if (!response.ok) {
-        throw new Error('Network response was not ok');
-      }
-
-      const blob = await response.blob();
-      const url = window.URL.createObjectURL(blob);
-      const link = document.createElement('a');
-      link.href = url;
-      link.download = 'AdminManual.pdf';
-      document.body.appendChild(link);
-      link.click();
-      document.body.removeChild(link);
-      window.URL.revokeObjectURL(url);
-    } catch (error) {
-      console.error('There has been a problem with your fetch operation:', error);
-    }
-  } else if (store.getters.userRole === 'ROLE_ADMIN') {
     try {
       const response = await fetch('http://localhost:5000/admin/pdfdownload/admin-pdf', {
         method: 'GET',
@@ -239,7 +212,6 @@ const pdfDownload = async () => {
       console.error('There has been a problem with your fetch operation:', error);
     }
   }
-};
 
 const logout = async () => {
   try {
